Implement own IRC watchdog to detect network disconnections
The goirc's own ping loop (configured with `PingFreq`) only sends PINGs
to the server but doesn't handle failures to detect disconnections, so
it would take about 15 minutes (on Linux) to detect and force a
reconnection, which is unwanted.  This is supposed to be caused by the
TCP keepalive of the underlying socket.

Without being able to disable TCP keepalive using goirc's API, implement
our own watchdog that uses PINGs to proactively detect failures and
trigger reconnections.

Credits: ChatGPT and Claude Sonnet